Move newly added test and benchmark to jetstream_consumer_test.go
