+++ server/jetstream_consumer_test.go
@@ -21,12 +21,14 @@ import (
 	"fmt"
 	"math/rand"
 	"sort"
+	"strings"
 	"sync"
 	"sync/atomic"
 	"testing"
 	"time"
 
 	"github.com/nats-io/nats.go"
+	"github.com/nats-io/nuid"
 )
 
 func TestJetStreamConsumerMultipleFiltersRemoveFilters(t *testing.T) {
@@ -791,4 +793,91 @@ func TestJetStreamConsumerMultipleFiltersLastPerSubject(t *testing.T) {
 	require_Equal(t, "3", string(msg.Data))
 	require_Equal(t, "two", msg.Subject)
 
+}
+
+func consumerWithFilterSubjects(filterSubjects []string) *consumer {
+	c := consumer{}
+	for _, filter := range filterSubjects {
+		sub := &subjectFilter{
+			subject:          filter,
+			hasWildcard:      subjectHasWildcard(filter),
+			tokenizedSubject: tokenizeSubjectIntoSlice(nil, filter),
+		}
+		c.subjf = append(c.subjf, sub)
+	}
+
+	return &c
+}
+
+func filterSubjects(n int) []string {
+	fs := make([]string, 0, n)
+	for {
+		literals := []string{"foo", "bar", nuid.Next(), "xyz", "abcdef"}
+		fs = append(fs, strings.Join(literals, "."))
+		if len(fs) == n {
+			return fs
+		}
+		// Create more filterSubjects by going through the literals and replacing one with the '*' wildcard.
+		l := len(literals)
+		for i := 0; i < l; i++ {
+			e := make([]string, l)
+			for j := 0; j < l; j++ {
+				if j == i {
+					e[j] = "*"
+				} else {
+					e[j] = literals[j]
+				}
+			}
+			fs = append(fs, strings.Join(e, "."))
+			if len(fs) == n {
+				return fs
+			}
+		}
+	}
+}
+
+func TestJetStreamConsumerIsFilteredMatch(t *testing.T) {
+	for _, test := range []struct {
+		name           string
+		filterSubjects []string
+		subject        string
+		result         bool
+	}{
+		{"no filter", []string{}, "foo.bar", true},
+		{"literal match", []string{"foo.baz", "foo.bar"}, "foo.bar", true},
+		{"literal mismatch", []string{"foo.baz", "foo.bar"}, "foo.ban", false},
+		{"wildcard > match", []string{"bar.>", "foo.>"}, "foo.bar", true},
+		{"wildcard > match", []string{"bar.>", "foo.>"}, "bar.foo", true},
+		{"wildcard > mismatch", []string{"bar.>", "foo.>"}, "baz.foo", false},
+		{"wildcard * match", []string{"bar.*", "foo.*"}, "foo.bar", true},
+		{"wildcard * match", []string{"bar.*", "foo.*"}, "bar.foo", true},
+		{"wildcard * mismatch", []string{"bar.*", "foo.*"}, "baz.foo", false},
+		{"wildcard * match", []string{"foo.*.x", "foo.*.y"}, "foo.bar.x", true},
+		{"wildcard * match", []string{"foo.*.x", "foo.*.y", "foo.*.z"}, "foo.bar.z", true},
+		{"many mismatch", filterSubjects(100), "foo.bar.do.not.match.any.filter.subject", false},
+		{"many match", filterSubjects(100), "foo.bar.12345.xyz.abcdef", true}, // will be matched by "foo.bar.*.xyz.abcdef"
+	} {
+		test := test
+
+		t.Run(test.name, func(t *testing.T) {
+			t.Parallel()
+
+			c := consumerWithFilterSubjects(test.filterSubjects)
+			if res := c.isFilteredMatch(test.subject); res != test.result {
+				t.Fatalf("Subject %q filtered match of %v, should be %v, got %v",
+					test.subject, test.filterSubjects, test.result, res)
+			}
+		})
+	}
+}
+
+func Benchmark____JetStreamConsumerIsFilteredMatch(b *testing.B) {
+	subject := "foo.bar.do.not.match.any.filter.subject"
+	for n := 1; n <= 1024; n *= 2 {
+		name := fmt.Sprintf("%d filter subjects", int(n))
+		c := consumerWithFilterSubjects(filterSubjects(int(n)))
+		b.Run(name, func(b *testing.B) {
+			c.isFilteredMatch(subject)
+		})
+	}
 }
